+++ test/Main.hs
@@ -1124,9 +1124,8 @@ bankTrigger1 =
   Trigger
     { triggerTable = "bank"
     , triggerName = "trigger_1"
+    , triggerKind = TriggerConstraint NotDeferrable
     , triggerEvents = Set.fromList [TriggerInsert]
-    , triggerDeferrable = False
-    , triggerInitiallyDeferred = False
     , triggerWhen = Nothing
     , triggerFunction =
         "begin"
@@ -1149,9 +1148,8 @@ bankTrigger3 =
   Trigger
     { triggerTable = "bank"
     , triggerName = "trigger_3"
+    , triggerKind = TriggerConstraint DeferrableInitiallyDeferred
     , triggerEvents = Set.fromList [TriggerInsert, TriggerUpdateOf [unsafeSQL "location"]]
-    , triggerDeferrable = True
-    , triggerInitiallyDeferred = True
     , triggerWhen = Nothing
     , triggerFunction =
         "begin"
@@ -1318,8 +1316,8 @@ testTriggers step = do
       verify [trg] True
 
   do
-    let msg = "successfully migrate trigger that is deferrable"
-        trg = bankTrigger1 {triggerDeferrable = True}
+    let msg = "successfully migrate a constraint trigger that is deferrable"
+        trg = bankTrigger1 {triggerKind = TriggerConstraint Deferrable}
         ts =
           [ tableBankSchema1
               { tblVersion = 2
@@ -1332,12 +1330,8 @@ testTriggers step = do
       verify [trg] True
 
   do
-    let msg = "successfully migrate trigger that is deferrable and initially deferred"
-        trg =
-          bankTrigger1
-            { triggerDeferrable = True
-            , triggerInitiallyDeferred = True
-            }
+    let msg = "successfully migrate a constraint trigger that is deferrable and initially deferred"
+        trg = bankTrigger1 {triggerKind = TriggerConstraint DeferrableInitiallyDeferred}
         ts =
           [ tableBankSchema1
               { tblVersion = 2
@@ -1350,12 +1344,8 @@ testTriggers step = do
       verify [trg] True
 
   do
-    let msg = "database exception is raised if trigger is initially deferred but not deferrable"
-        trg =
-          bankTrigger1
-            { triggerDeferrable = False
-            , triggerInitiallyDeferred = True
-            }
+    let msg = "successfully migrate a regular AFTER trigger"
+        trg = bankTrigger1 {triggerKind = TriggerRegular After}
         ts =
           [ tableBankSchema1
               { tblVersion = 2
@@ -1364,7 +1354,22 @@ testTriggers step = do
           ]
         ms = [createTriggerMigration 1 trg]
     triggerStep msg $ do
-      assertDBException msg $ migrate ts ms
+      assertNoException msg $ migrate ts ms
+      verify [trg] True
+
+  do
+    let msg = "successfully migrate a regular BEFORE trigger"
+        trg = bankTrigger1 {triggerKind = TriggerRegular Before}
+        ts =
+          [ tableBankSchema1
+              { tblVersion = 2
+              , tblTriggers = [trg]
+              }
+          ]
+        ms = [createTriggerMigration 1 trg]
+    triggerStep msg $ do
+      assertNoException msg $ migrate ts ms
+      verify [trg] True
 
   do
     let msg = "database exception is raised if dropping trigger that does not exist"